updated docs; reorganized modules
