+++ astroqtpy/base_node.py
@@ -1,17 +1,57 @@
 import abc
+from dataclasses import dataclass
+
 import matplotlib as mpl
 
+
+@dataclass
+class QuadPoint():
+    """Quadtree point.
+
+    A dataclass for storing a value at coordinates (x, y).
+    
+    Args:
+        _x (float): x position.
+        _y (float): y position.
+        _value (float, optional): Point value. Defaults to -1.
+    """
+    _x : float
+    _y : float
+    _value : float = -1
+    
+    @property
+    def x(self) -> float:
+        return self._x
+    @x.setter
+    def x(self, val: float) -> None:
+        self._x = val
+        
+    @property
+    def y(self) -> float:
+        return self._y
+    @y.setter
+    def y(self, val: float) -> None:
+        self._y = val
+        
+    @property
+    def value(self) -> float:
+        return self._value
+    @value.setter
+    def value(self, val: float) -> None:
+        self._value = val
+
+
 class QuadNode(abc.ABC):
-    """Quadtree node
+    """Quadtree base node.
 
     A base class for all released types of astroQTpy quadtree nodes.
     
     Args:
-            x_min (float): Minimum x value of this node.
-            x_max (float): Maximum x value of this node.
-            y_min (float): Minimum y value of this node.
-            y_max (float): Maximum y value of this node.
-            depth (int, optional): Depth of this node. Defaults to 0.
+        x_min (float): Minimum x value of this node.
+        x_max (float): Maximum x value of this node.
+        y_min (float): Minimum y value of this node.
+        y_max (float): Maximum y value of this node.
+        depth (int, optional): Depth of this node. Defaults to 0.
     """
     def __init__(self,
         x_min : float,
