+++ docs/source/nodes.rst
@@ -1,0 +1,9 @@
+.. _nodes:
+
+Nodes
+======
+
+Node classes to inherit the base node in astroQTpy.
+
+.. automodule:: astroqtpy.nodes
+   :members:
